Fixed typecasting

+++ com0com/sys/wmi.c
@@ -19,6 +19,10 @@
  *
  *
  * $Log$
+ * Revision 1.2  2006/10/17 06:54:37  vfrolov
+ * Disabled SERIAL_PORT_WMI_HW_GUID for binary compatibility with
+ * both W2K and WXP
+ *
  * Revision 1.1  2006/08/23 13:09:55  vfrolov
  * Initial revision
  *
@@ -113,7 +117,7 @@ NTSTATUS QueryWmiDataBlock(
   case COC_WMI_PORT_NAME:
     Trace0((PC0C_COMMON_EXTENSION)pDevExt, L"QueryWmiDataBlock PORT_NAME");
 
-    size = wcslen(pDevExt->portName) * sizeof(pDevExt->portName[0]);
+    size = (ULONG)wcslen(pDevExt->portName) * sizeof(pDevExt->portName[0]);
 
     if (bufSize < (size + sizeof(USHORT))) {
       size += sizeof(USHORT);
